multiple components sync and lifting state up
