+++ react-app/src/components/counters.jsx
@@ -2,41 +2,13 @@ import React, { Component } from 'react';
 import Counter from './counter';
 
 class Counters extends Component{
-    state = {
-        counters: [
-            {id: 0, value: 4},
-            {id: 1, value: 0},
-            {id: 2, value: 0},
-            {id: 3, value: 0}
-        ]
-    };
-
-    handleIncrement = counter => {
-        const counters = [...this.state.counters];
-        // console.log("counters array: ",JSON.stringify(counters));
-        const index = counters.indexOf(counter);
-        // console.log("index: ",index)
-        counters[index].value ++;
-        this.setState({counters});
-    };
-
-    handleDelete = counterId =>{
-        const counters = this.state.counters.filter(c => c.id !== counterId);
-        this.setState({counters});
-        console.log("event handlers %s %s", counterId,counters)
-    }
-
-    handleReset = () => {
-       const counters = this.state.counters.map(c => {c.value = 0; return c;});
-       this.setState({counters});
-    }
 
     render(){
         return (
             <div>
-                <button onClick={this.handleReset} className="btn btn-primary btn-sm m-2">Reset</button>
-                {this.state.counters.map(counter => 
-                <Counter key={counter.id} onIncrement={this.handleIncrement} onDelete={this.handleDelete} counter={counter}>
+                <button onClick={this.props.onReset} className="btn btn-primary btn-sm m-2">Reset</button>
+                {this.props.counters.map(counter => 
+                <Counter key={counter.id} onIncrement={this.props.onIncrement} onDelete={this.props.onDelete} counter={counter}>
                 <h4>Counter {counter.id}</h4>    
                 </Counter>
                 )}
